chore: se migra evaluacion360 a su propio modelo
