+++ talent360/evaluaciones/models/__init__.py
@@ -9,3 +9,4 @@ from . import respuesta
 from . import template
 from . import res_users
 from . import opcion
+from . import evaluacion_360
