+++ talent360/evaluaciones/security/ir.model.access.csv
@@ -8,3 +8,4 @@
 "access_template","access.template","model_template","base.group_user",1,1,1,1
 "access_usuario_evaluacion_rel","access.usuario.evaluacion.rel","model_usuario_evaluacion_rel","base.group_user",1,1,1,1
 "access_opcion","access.opcion","model_opcion","base.group_user",1,1,1,1
+"access_evaluacion_360","access.evaluacion.360","model_evaluacion_360","base.group_user",1,1,1,1
